bot/requirements.txt
requests
beautifulsoup4
gspread
oauth2client
python-dotenv
schedule
selenium
webdriver-manager
